Refactored module names (files) to be `snake_case`
To avoid confusion with class names, which are CamelCase
