+++ ScanAnalysis/multiscan_analysis/Undulator/multiscan_analysis.py
@@ -21,7 +21,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 from geecs_python_api.analysis.scans.scan_data import ScanData
-from scan_analysis.analyzers.Undulator.VisaEBeamAnalysis import VisaEBeamAnalysis
+from scan_analysis.analyzers.Undulator.visa_ebeam_analysis import VisaEBeamAnalysis
 from image_analysis.utils import read_imaq_image
 
 # =============================================================================
