+++ ScanAnalysis/scan_analysis/analyzers/Undulator/__init__.py
@@ -1,5 +1,5 @@
-from .HiResMagCamAnalysis import HiResMagCamAnalysis
-from .MagSpecStitcherAnalysis import MagSpecStitcherAnalysis
-from .Rad2SpecAnalysis import Rad2SpecAnalysis
-from .VisaEBeamAnalysis import VisaEBeamAnalysis
-from .CameraImageAnalysis import CameraImageAnalysis
+from .hi_res_mag_cam_analysis import HiResMagCamAnalysis
+from .mag_spec_stitcher_analysis import MagSpecStitcherAnalysis
+from .rad2_spec_analysis import Rad2SpecAnalysis
+from .visa_ebeam_analysis import VisaEBeamAnalysis
+from .camera_image_analysis import CameraImageAnalysis
