+++ ScanAnalysis/scan_analysis/analyzers/Undulator/visa_ebeam_analysis.py
@@ -12,7 +12,7 @@ if TYPE_CHECKING:
     from geecs_python_api.controls.api_defs import ScanTag
 import numpy as np
 import cv2
-from scan_analysis.analyzers.Undulator.CameraImageAnalysis import CameraImageAnalysis
+from scan_analysis.analyzers.Undulator.camera_image_analysis import CameraImageAnalysis
 from geecs_python_api.analysis.scans.scan_data import ScanData
